Update README.md
You are going to design an application to allow a user to

self-study using flash cards. In this part of the project,

a user will…

1. Be prompted to choose from a menu of available flash

card decks; this menu will repeat until a valid

selection is made.

2. Proceed through each card in the selected deck,

one-by-one. For each card, the front is displayed,

and the user is allowed time to reflect; then the

back is displayed; and the user is asked if they

got the correct answer.

3. Once the deck is exhausted, the program outputs the

number of self-reported correct answers and ends.

Of course, we’ll design this program step-by-step, AND

you’ve already done pieces of this in homework!!